feat(misc): Add notification permission request for Battery Guru
This commit introduces a runtime permission request for `POST_NOTIFICATIONS` on Android 13 (TIRAMISU) and above

If the notification permission is not granted, a button is now displayed within the Battery Info card, allowing the user to grant it directly. The state of the permission is checked and updated using `rememberLauncherForActivityResult`.

The layout of the `BatteryInfoSection` has been adjusted from a `Row` to a `Column` to accommodate the new permission button.